Planificador corto plazo e interrupcion Finalizar

+++ entradasalida/include/auxiliares.h
@@ -12,7 +12,7 @@
 extern char* nombre_interfaz;
 
 void iniciar_hilo_kernel(t_config* config);
-void entender_mensajes(char* mensaje, int socket_cliente,double tiempo_unidad_trabajo);
-void io_gen_sleep(char* unidades_tiempo, double tiempo_unidad_trabajo, int socket_cliente);
+void entender_mensajes(char* mensaje, int socket_cliente,int tiempo_unidad_trabajo);
+void io_gen_sleep(char* unidades_tiempo, int tiempo_unidad_trabajo, int socket_cliente);
 
 #endif
